+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - SmartResize.Policy manual resize broken (https://github.com/edvin/tornadofx/issues/570)
 
 ### Changes
+- ItemViewModel.bind `defaultValue` parameter
 - Node builders inside of `MenuItem` will automatically assign the node to the `graphic` property of the menu item
 - The App class (main application entrypoint) no longer requires a primary view parameter, in case you want to show a tray icon or determinine what view to show some other way