+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 - SmartResize.Policy manual resize broken (https://github.com/edvin/tornadofx/issues/570)
 - TableView bound to ListProperty should rebind when value changes
+- Allow calling Workspace.disableNavigation() when workspace is empty
 
 ### Changes
 - AnchorPaneConstraint properties now accept any Number, not just Double